fix: file not found

+++ playwright/clean-x.js
@@ -7,8 +7,10 @@ const exityp = '"exit_type":"'
 module.exports = () => {
   const {home} = global.mitm.path
   const path = `${home}${profile}`
-  const oldstate1 = fs.readFileSync(path)
-  const newstate1 = `${oldstate1}`.replace(`${exited}false`  , `${exited}true`)
-  const newstate2 = `${newstate1}`.replace(`${exityp}Crashed`, `${exityp}none`)
-  fs.writeFileSync(path, newstate2)
+  if (fs.existsSync(path)) {
+    const oldstate1 = fs.readFileSync(path)
+    const newstate1 = `${oldstate1}`.replace(`${exited}false`  , `${exited}true`)
+    const newstate2 = `${newstate1}`.replace(`${exityp}Crashed`, `${exityp}none`)
+    fs.writeFileSync(path, newstate2)
+  }
 }
